+++ lib/xlsxir.ex
@@ -1,45 +1,36 @@
 defmodule Xlsxir do
-  require IEx
   alias Xlsxir.{Unzip, SaxParser, Worksheet, Timer}
 
   @moduledoc """
-  Extracts and parses data from a Microsoft Excel workbook, returning either a list or a map.
+  Extracts and parses data from a `.xlsx` file to an Erlang Term Storage (ETS) table and provides various functions for accessing the data.
   """
 
   @doc """
-  Provides a list of row value lists or a map of cell/value pairs from a Microsoft Excel workbook
-  given the path to a file of extension type `.xlsx`, the index of the worksheet in which the information
-  is requested, and an option.
+  Extracts worksheet data contained in the specified `.xlsx` file to an ETS table named `:worksheet` which is accessed via the `Xlsxir.Worksheet` module. Successful extraction 
+  returns `:ok` with the timer argument set to false and returns a tuple of `{:ok, time}` where time is a list containing time elapsed during the extraction process 
+  (i.e. `[hour, minute, second, microsecond]`) when the timer argument is set to true. 
 
-  Cells containing formulas return either a `string`, `integer` or `float` of the resulting value. Cells containing Excel date format return
-  a date in Erlang `:calendar.date()` format (i.e. `{year, month, day}`).
+  Cells containing formulas in the worksheet are extracted as either a `string`, `integer` or `float` depending on the resulting value of the cell. 
+  Cells containing an ISO 8601 date format are extracted and converted to Erlang `:calendar.date()` format (i.e. `{year, month, day}`).
 
   ## Parameters
   - `path` - file path of a `.xlsx` file type in `string` format
-  - `index` - index of worksheet from within the Excel workbook to be parsed, starting with `0`
-  - `option` - one of two available options
-
-  ## Options
-
-  - `:rows` - a list of row value lists (default) - i.e. [[row_1_values], [row_2_values], ...]
-  - `:cells` - a map of cell/value pairs - i.e. %{ A1: value_of_cell, B1: value_of_cell, ...}
+  - `index` - index of worksheet from within the Excel workbook to be parsed (zero-based index)
+  - `timer` - boolean flag that tracts extraction process time and returns it when set to `true`. Defalut value is `false`.
 
   ## Example
-    An example file named `test.xlsx` located in `./test/test_data` containing the following:
-    - cell 'A1' -> "string one"
-    - cell 'B1' -> "string two"
-    - cell 'C1' -> integer of 10
-    - cell 'D1' -> formula of "4 * 5"
-    - cell 'E1' -> date of 1/1/2016 or Excel date serial of 42370
+  Extract first worksheet in an example file named `test.xlsx` located in `./test/test_data`:
 
-          iex> Xlsxir.extract("./test/test_data/test.xlsx", 0)
-          [["string one", "string two", 10, 20, {2016, 1, 1}]]
-
-          iex> Xlsxir.extract("./test/test_data/test.xlsx", 0, :cells)
-          %{ A1: "string one", B1: "string two", C1: 10, D1: 20, E1: {2016,1,1}}
+      iex> Xlsxir.extract("./test/test_data/test.xlsx", 0)
+      :ok
+      iex> Xlsxir.Worksheet.alive?
+      true
+      iex> Xlsxir.close
+      :ok
   """
-  def extract(path, index) do
-    Timer.start
+  def extract(path, index, timer \\ false) do
+    if timer, do: Timer.start
+
     {:ok, file}       = Unzip.validate_path(path)
     {:ok, file_paths} = Unzip.xml_file_list(index)
                         |> Unzip.extract_xml_to_file(file)
@@ -48,14 +39,37 @@ defmodule Xlsxir do
       case file do
         'temp/xl/sharedStrings.xml' -> SaxParser.parse(to_string(file), :string)
         'temp/xl/styles.xml'        -> SaxParser.parse(to_string(file), :style)
-        _                           -> SaxParser.parse(to_string(file), :worksheet)
+        _                           -> nil
       end
     end)
 
-    Unzip.delete_temp_dir
-    {:ok, Timer.stop}
+    SaxParser.parse("temp/xl/worksheets/sheet#{index + 1}.xml", :worksheet)
+    Unzip.delete_dir
+
+    case timer do
+      false -> :ok
+      true  -> {:ok, Timer.stop}
+    end
   end
 
+  @doc """
+  Accesses `:worksheet` ETS table and returns data formatted as a list of row value lists.
+
+  ## Example
+  An example file named `test.xlsx` located in `./test/test_data` containing the following:
+  - cell 'A1' -> "string one"
+  - cell 'B1' -> "string two"
+  - cell 'C1' -> integer of 10
+  - cell 'D1' -> formula of "4 * 5"
+  - cell 'E1' -> date of 1/1/2016 or Excel date serial of 42370
+
+        iex> Xlsxir.extract("./test/test_data/test.xlsx", 0)
+        :ok
+        iex> Xlsxir.get_list
+        [["string one", "string two", 10, 20, {2016, 1, 1}]]
+        iex> Xlsxir.close
+        :ok
+  """
   def get_list do
     range = 0..(:ets.info(:worksheet, :size) -1)
 
@@ -65,6 +79,24 @@ defmodule Xlsxir do
                       end)
   end
 
+  @doc """
+  Accesses `:worksheet` ETS table and returns data formatted as a map of cell references and values.
+
+  ## Example
+  An example file named `test.xlsx` located in `./test/test_data` containing the following:
+  - cell 'A1' -> "string one"
+  - cell 'B1' -> "string two"
+  - cell 'C1' -> integer of 10
+  - cell 'D1' -> formula of "4 * 5"
+  - cell 'E1' -> date of 1/1/2016 or Excel date serial of 42370
+
+        iex> Xlsxir.extract("./test/test_data/test.xlsx", 0)
+        :ok
+        iex> Xlsxir.get_map
+        %{ "A1" => "string one", "B1" => "string two", "C1" => 10, "D1" => 20, "E1" => {2016,1,1}}
+        iex> Xlsxir.close
+        :ok
+  """
   def get_map do
     range = 0..(:ets.info(:worksheet, :size) -1)
 
@@ -75,32 +107,101 @@ defmodule Xlsxir do
                                  end)
   end
 
+  @doc """
+  Accesses `:worksheet` ETS table and returns value of specified cell. Note: entire worksheet is traversed each time this function is executed. Use with caution as 
+  this would be an expensive task for a large worksheet.
+
+  ## Parameters
+  - `cell_ref` - Reference name of cell to be returned in `string` format (i.e. `"A1"`)
+
+  ## Example
+  An example file named `test.xlsx` located in `./test/test_data` containing the following:
+  - cell 'A1' -> "string one"
+  - cell 'B1' -> "string two"
+  - cell 'C1' -> integer of 10
+  - cell 'D1' -> formula of "4 * 5"
+  - cell 'E1' -> date of 1/1/2016 or Excel date serial of 42370
+
+        iex> Xlsxir.extract("./test/test_data/test.xlsx", 0)
+        :ok
+        iex> Xlsxir.get_cell("A1")
+        "string one"
+        iex> Xlsxir.close
+        :ok
+  """
   def get_cell(cell_ref) do
-    default = "No value found in cell #{cell_ref}"
-
-    row = ~r/\d+/
-          |> Regex.scan(cell_ref)
-          |> List.flatten
-          |> List.first
-          |> String.to_integer
-
-    [_, value] = Enum.find(Worksheet.get_at(row - 1), default, fn cell -> 
-                   Enum.at(cell, 0) == cell_ref 
-                 end)
-    value
+    data = get_map
+    data[cell_ref]
   end
 
+  @doc """
+  Accesses `:worksheet` ETS table and returns values of specified row in a `list`. Blank rows are not counted.
+
+  ## Parameters
+  - `row` - Reference name of row to be returned in `integer` format (i.e. `1`)
+
+  ## Example
+  An example file named `test.xlsx` located in `./test/test_data` containing the following:
+  - cell 'A1' -> "string one"
+  - cell 'B1' -> "string two"
+  - cell 'C1' -> integer of 10
+  - cell 'D1' -> formula of "4 * 5"
+  - cell 'E1' -> date of 1/1/2016 or Excel date serial of 42370
+
+        iex> Xlsxir.extract("./test/test_data/test.xlsx", 0)
+        :ok
+        iex> Xlsxir.get_row(1)
+        ["string one", "string two", 10, 20, {2016, 1, 1}]
+        iex> Xlsxir.close
+        :ok
+  """
   def get_row(row) do
-    Enum.map(Worksheet.get_at(row - 1), fn [k,v] -> v end)
+    Enum.map(Worksheet.get_at(row - 1), fn [_k, v] -> v end)
   end
 
+  @doc """
+  Accesses `:worksheet` ETS table and returns values of specified column in a `list`.
+
+  ## Parameters
+  - `col` - Reference name of column to be returned in `string` format (i.e. `"A"`)
+
+  ## Example
+  An example file named `test.xlsx` located in `./test/test_data` containing the following:
+  - cell 'A1' -> "string one"
+  - cell 'B1' -> "string two"
+  - cell 'C1' -> integer of 10
+  - cell 'D1' -> formula of "4 * 5"
+  - cell 'E1' -> date of 1/1/2016 or Excel date serial of 42370
+
+        iex> Xlsxir.extract("./test/test_data/test.xlsx", 0)
+        :ok
+        iex> Xlsxir.get_col("A")
+        ["string one"]
+        iex> Xlsxir.close
+        :ok
+  """
   def get_col(col) do
     Enum.map(get_map, fn {k,v} -> if cell_ltrs(k) == col, do: v end)
     |> Enum.reject(fn x -> x == nil end)
   end
 
+  @doc """
+  Deletes ETS table `:worksheet` and returns `:ok` if successful.
+
+  ## Example
+  Extract first worksheet in an example file named `test.xlsx` located in `./test/test_data`:
+
+      iex> Xlsxir.extract("./test/test_data/test.xlsx", 0)
+      :ok
+      iex> Xlsxir.close
+      :ok
+  """
   def close do
     Worksheet.delete
+    |> case do
+      false -> raise "Unable to close worksheet"
+      true -> :ok
+    end
   end
 
   defp cell_ltrs(cell) do