+++ lib/xlsxir/parse_string.ex
@@ -11,17 +11,17 @@ defmodule Xlsxir.ParseString do
 
   @doc """
   Sax event utilized by `Xlsxir.SaxParser.parse/2`. Takes a pattern and the current state of a struct and recursivly parses the
-  sharedString XML file, ultimately sending each parsed string to the `SharedString` agent process that was started by 
+  sharedString XML file, ultimately sending each parsed string to the `Xlsxir.SharedString` module which contains an ETS table opened by 
   `Xlsxir.SaxParser.parse/2`. 
 
   ## Parameters
 
-  - pattern - the XML pattern of the event to match upon
-  - state - the state argument is unused when parsing strings and is therefore proceeded by an underscore
+  - pattern - the XML pattern of the event to match on
+  - state - current state of the `%StringState{}` struct
 
   ## Example
-  Recursively sends strings from the `xl/sharedStrings.xml` file to `SharedString.add_shared_string/1`. The data can ultimately
-  be retreived by the `get/0` function of the agent process (i.e. `Xlsxir.SharedString.get` would return `["string 1", "string 2", ...]`).
+  Recursively sends strings from the `xl/sharedStrings.xml` file to `Xlsxir.SharedString.add_shared_string/2`. The data can ultimately
+  be retreived by the `get_at/1` function of the `Xlsxir.SharedString` module (i.e. `Xlsxir.SharedString.get_at(0)` would return something like `"string 1"`).
   """
   def sax_event_handler(:startDocument, _state), do: Index.new
 
@@ -36,7 +36,7 @@ defmodule Xlsxir.ParseString do
     %{state | empty_string: false}
   end
 
-  def sax_event_handler({:endElement,_,'si',_}, %StringState{empty_string: empty_string} = state) do
+  def sax_event_handler({:endElement,_,'si',_}, %StringState{empty_string: empty_string}) do
     if empty_string do 
       SharedString.add_shared_string("", Index.get)
       Index.increment_1